+++ R/traits.R
@@ -140,3 +140,272 @@ map_trait_distribution <- function(tidy_trait,
     )
   return(trait_map)
 }
+
+# `ggplot` Wrappers ----
+
+#' Built violin plots with jittered points
+#'
+#' Semi-flexible plotting of continuous trait violin plots with adjustable
+#' point jitters.
+#'
+#' @param trait Variable with trait observation in
+#'   [ThesisPackage::herbarium_specimens].
+#' @param aesthetic_id Variable to set color and shape aesthetics of jitters.
+#' @param aesthetic_labels Named vector of labels for manual scale values.
+#' @param violin.params List of arguments passed to [ggplot2::geom_violin().
+#' @param jitter.params List of arguments passed to [ggplot2::geom_jitter()].
+#' @param theme.params List of arguments passed to [ggplot2::theme()].
+#' @inheritParams layer_specimens
+#' @inheritParams layer_themes
+#' @export
+#'
+#' @return [ggplot] of trait observations delineated by reviewed id (x-axis)
+#'   and jittered points (y-axis) representing numeric values.
+#'
+jitter_violin <- function(specimen_tbl, trait, aesthetic_id,
+                          aesthetic_labels = NULL, legend_title = NULL,
+                          violin.params = list(),
+                          jitter.params = list(),
+                          theme.params = list()) {
+
+  # Text markdown label replacements to legend text from prior & reviewed IDs.
+  if (is.null(aesthetic_labels)) {
+    aesthetic_labels <-
+      ThesisPackage::spl_labels(
+        specimen_tbl = specimen_tbl,
+        id_column = aesthetic_id
+      )
+  }
+
+  violins <- do.call("geom_violin", utils::modifyList(
+    list(
+      mapping = aes_(x = quote(Taxon_a_posteriori), y = as.name(trait)),
+      na.rm = TRUE),
+    val = violin.params)
+  )
+
+  jitters <- do.call("geom_jitter", utils::modifyList(
+    list(
+      mapping = aes_(
+        x = quote(Taxon_a_posteriori),
+        y = as.name(trait),
+        color = as.name(aesthetic_id),
+        shape = as.name(aesthetic_id)
+        ),
+      na.rm = TRUE
+    ),
+    val = jitter.params)
+  )
+
+  themes <- do.call("theme", utils::modifyList(
+    list(
+      title = ggtext::element_markdown(),
+      axis.title.x = ggplot2::element_blank(),
+      legend.position = "none"
+    ),
+    val = theme.params)
+  )
+
+  ggplot(data = specimen_tbl) +
+    ggplot2::theme_classic() +
+    violins +
+    jitters +
+    ggplot2::scale_color_manual(
+      name = legend_title, labels = aesthetic_labels,
+      values = ThesisPackage::spp_color, na.value = "black"
+    ) +
+    ggplot2::scale_shape_manual(
+      name = legend_title, labels = aesthetic_labels,
+      values = ThesisPackage::spp_shape, na.value = 17
+    ) +
+    themes
+}
+
+#' Plot Trait Phenology
+#'
+#' Wrapper to [ggplot()] layering points by collection date for a given
+#' y variable speecified by `trait`. The `aesthetic_id` determines
+#' **color** and **shape** geom point aesthetics and `aesthetic_labels`
+#' are passed as manual scale specifications. Use `legend_title` to set
+#' the combined color and shape scale legend title.
+#'
+#' Additional paramaters can be passed to `geom_jitter()` via `jitter.params`
+#' and [theme()] via `theme.params()`, where each `\*.params()` argument
+#' takes a list
+#'
+#' @inheritParams jitter_violin
+#' @inheritParams layer_themes
+#' @importFrom ggplot2 geom_point theme
+#' @export
+#'
+#' @return [ggplot] object of trait phenology by "%m-%d" date format.
+#'
+#' @examples
+#' herbarium_specimens %>%
+#'   subset_coords(Longitude = c(-108, -105), Latitude = c(39, 42)) %>%
+#'   dplyr::bind_cols(.,
+#'     range_split(trait_tbl = ., split_var = "Mature_fruit_length_mm")
+#'   ) %>%
+#'   trait_phenology(
+#'     specimen_tbl = .,
+#'     trait = "Mature_fruit_length_mm_max",
+#'     aesthetic_id = "Taxon_a_posteriori",
+#'     legend_title = "Reviewed Annotations"
+#'   )
+#'
+trait_phenology <- function(specimen_tbl, trait, aesthetic_id,
+                            aesthetic_labels = NULL, legend_title = NULL,
+                            jitter.params = list(),
+                            theme.params = list()) {
+
+  specimen_tbl <- dplyr::filter(specimen_tbl,
+                                !is.na(.data$Date_md) & !is.na(.data[[trait]]))
+
+  # Text markdown label replacements to legend text from prior & reviewed IDs.
+  if (is.null(aesthetic_labels)) {
+    aesthetic_labels <-
+      ThesisPackage::spl_labels(
+        specimen_tbl = specimen_tbl,
+        id_column = aesthetic_id
+      )
+  }
+
+  jitters <- do.call("geom_point", utils::modifyList(
+    list(na.rm = TRUE),
+    val = jitter.params)
+  )
+
+  scales <- list(
+    ggplot2::scale_x_date(
+      name = "Collection Date",
+      date_labels = "%b",
+      limits = as.Date(c("05-01", "08-31"), format = "%m-%d")
+    ),
+    ggplot2::scale_color_manual(
+      name = legend_title,
+      labels = aesthetic_labels,
+      values = ThesisPackage::spp_color,
+      na.value = "black"
+    ),
+    ggplot2::scale_shape_manual(
+      name = legend_title,
+      labels = aesthetic_labels,
+      values = ThesisPackage::spp_shape,
+      na.value = 17
+    )
+  )
+
+  themes <- do.call("theme", utils::modifyList(
+    list(legend.text = ggtext::element_markdown()),
+    val = theme.params)
+  )
+
+  ggplot2::ggplot(
+    data = specimen_tbl,
+    mapping = aes_(
+      x = quote(Date_md),
+      y = as.name(trait),
+      color = as.name(aesthetic_id),
+      shape = as.name(aesthetic_id)
+      )
+    ) +
+    ggplot2::geom_smooth(method = 'lm', na.rm = TRUE, formula = y ~ x) +
+    ggplot2::theme_classic() +
+    jitters +
+    scales +
+    themes
+}
+
+#' Extract Legend of [ggplot] Aesthetics from Specimen Annotations
+#'
+#' @param ncol Number of legend columns.
+#' @inheritParams jitter_violin
+#' @inheritParams layer_specimens
+#' @inheritParams layer_themes
+#' @importFrom ggplot2 theme
+#' @export
+#'
+#' @return A `grob` with extracted plot legend of annotation aesthetics.
+#'
+#' @examples
+#' specimen_tbl <- herbarium_specimens %>%
+#'   subset_coords(Longitude = c(-108, -105), Latitude = c(39, 42))
+#' aesthetic_id <- "Taxon_a_posteriori"
+#' legend_title <- "Reviewed Annotations"
+#' aesthetic_labels <- NULL
+#' theme.params <- list()
+#' guide.params <- list()
+#'
+annotation_legend <- function(specimen_tbl, aesthetic_id, ncol = 2,
+                              legend_title, aesthetic_labels = NULL,
+                              theme.params = list()) {
+
+  # Text markdown label replacements to legend text from prior & reviewed IDs.
+  if (is.null(aesthetic_labels)) {
+    aesthetic_labels <-
+      ThesisPackage::spl_labels(
+        specimen_tbl = specimen_tbl,
+        id_column = aesthetic_id
+      )
+  }
+
+  plot_themes <- do.call("theme", utils::modifyList(
+    list(
+      title = ggtext::element_markdown(),
+      axis.text = ggplot2::element_text(angle = 45, hjust = 1),
+      legend.title = ggplot2::element_text(colour = "grey"),
+      legend.title.align = 0.5,
+      legend.justification = "center",
+      legend.position = "bottom",
+      legend.text = ggtext::element_markdown()
+    ),
+    val = theme.params)
+  )
+
+  plot_layers <-
+    list(
+      geom_jitter(
+        aes_(
+          x = quote(Taxon_a_posteriori),
+          y = as.name(aesthetic_id),
+          color = as.name(aesthetic_id),
+          shape = as.name(aesthetic_id)
+        ),
+        width = 0.25, height = 0.25
+      ),
+      ggplot2::scale_color_manual(
+        name = legend_title,
+        labels = aesthetic_labels,
+        values = ThesisPackage::spp_color,
+        na.value = "black"
+      ),
+      ggplot2::scale_shape_manual(
+        name = legend_title,
+        labels = aesthetic_labels,
+        values = ThesisPackage::spp_shape,
+        na.value = 17
+      ),
+      ggplot2::guides(
+        "color" = ggplot2::guide_legend(
+          ncol = ncol,
+          byrow = TRUE,
+          title.position = "top"
+        ),
+        "shape" = ggplot2::guide_legend(
+          ncol = ncol,
+          byrow = TRUE,
+          title.position = "top"
+        )
+      ),
+      ggplot2::theme_classic(),
+      ggplot2::labs(x = "Reviewed", y = "Prior")
+    )
+
+  built_ggplot <- ggplot(data = specimen_tbl) +
+    plot_layers +
+    plot_themes
+
+  ggplot_legend <- cowplot::get_legend(built_ggplot)
+  return(ggplot_legend)
+}
+